+++ packages/components/src/Chips/ChipsTypes.tsx
@@ -8,7 +8,7 @@ interface ChipFoundationProps {
   /**
    * Change the interaction on the chip.
    */
-  readonly type?: "singleselect" | "multiselect";
+  readonly type?: "singleselect" | "multiselect" | "dismissible";
 
   /**
    * Determines which chip gets highlighted.
@@ -54,4 +54,49 @@ export interface ChipMultiSelectProps extends ChipFoundationProps {
   onChange(value: string[]): void;
 }
 
-export type ChipsProps = XOR<ChipSingleSelectProps, ChipMultiSelectProps>;
+export interface ChipDismissibleProps extends ChipFoundationProps {
+  readonly type: "dismissible";
+  readonly selected: string[];
+  onChange(value: string[]): void;
+
+  /**
+   * Use a custom activator to trigger the Chip option selector
+   */
+  readonly activator?: ReactElement;
+
+  /**
+   * Adds a loading indicator
+   */
+  readonly isLoadingMore?: boolean;
+
+  /**
+   * Callback when the user selects the custom option instead of the available
+   * chips.
+   *
+   * If not implemented, it won't allow custom add.
+   *
+   * @param value
+   */
+  onCustomAdd?(value: string): void;
+
+  /**
+   * Callback when a user types a word that filters the options. Use this when
+   * you need to query new options from the database based on the search value.
+   *
+   * @param searchValue - The input value
+   */
+  onSearch?(searchValue: string): void;
+
+  /**
+   * Callback when the user scrolls at the end of the chip option list. Use this
+   * to load more options from the database.
+   *
+   * @param searchValue - The input value
+   */
+  onLoadMore?(searchValue: string): void;
+}
+
+export type ChipsProps = XOR<
+  ChipSingleSelectProps,
+  XOR<ChipMultiSelectProps, ChipDismissibleProps>
+>;