+++ packages/components/src/Chips/InternalChipDismissible/index.ts
@@ -1,0 +1,1 @@
+export { InternalChipDismissible } from "./InternalChipDismissible";