+++ packages/design/src/index.ts
@@ -7,3 +7,4 @@ export {
   sizesClassMap,
   colorsClassMap,
 } from "./icons/getIcon";
+export { iconStyles } from "./icons/iconStyles";