+++ packages/design/CHANGELOG.md
@@ -8,6 +8,39 @@ menu: Changelog
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.24.0](https://github.com/GetJobber/atlantis/compare/@jobber/design@0.23.0...@jobber/design@0.24.0) (2022-03-03)
+
+
+### Features
+
+* **design:** update overlay color to black ([#861](https://github.com/GetJobber/atlantis/issues/861)) ([40351db](https://github.com/GetJobber/atlantis/commit/40351db7bf7b8f7494ca7d3d70b2a98ff46b1203))
+
+
+
+
+
+# [0.23.0](https://github.com/GetJobber/atlantis/compare/@jobber/design@0.22.0...@jobber/design@0.23.0) (2022-02-22)
+
+
+### Features
+
+* **design:** add new icons to atlantis -- (Minor) Feature Release ([#860](https://github.com/GetJobber/atlantis/issues/860)) ([feca432](https://github.com/GetJobber/atlantis/commit/feca432a9234d2849203f5edbed06c58839e3175))
+
+
+
+
+
+# [0.22.0](https://github.com/GetJobber/atlantis/compare/@jobber/design@0.21.0...@jobber/design@0.22.0) (2022-02-07)
+
+
+### Features
+
+* **design:** Build Icon CSS to a CSS-in-JS react native format ([#850](https://github.com/GetJobber/atlantis/issues/850)) ([5ff025f](https://github.com/GetJobber/atlantis/commit/5ff025ff8fd770f5c4919afa5fe5a0387c6bfe8a))
+
+
+
+
+
 # [0.21.0](https://github.com/GetJobber/atlantis/compare/@jobber/design@0.20.1...@jobber/design@0.21.0) (2021-12-18)